+++ components/wearable/emulate/common/src/main/java/com/flipperdevices/wearable/emulate/common/WearableCommandInputStream.kt
@@ -78,5 +78,4 @@ class WearableCommandInputStream<T : GeneratedMessageLite<*, *>>(
             delay(READ_TIMEOUT_MS)
         }
     }
-
 }