+++ components/wearable/emulate/handheld/impl/src/main/java/com/flipperdevices/wearable/emulate/handheld/impl/di/WearServiceComponent.kt
@@ -26,7 +26,6 @@ interface WearServiceComponentDependencies {
     val keyParser: KeyParser
 }
 
-
 @SingleIn(WearHandheldGraph::class)
 @MergeComponent(
     WearHandheldGraph::class,