+++ components/wearable/emulate/handheld/impl/src/main/java/com/flipperdevices/wearable/emulate/handheld/impl/request/WearablePingProcessor.kt
@@ -19,9 +19,11 @@ class WearablePingProcessor @Inject constructor(
     private val scope: CoroutineScope
 ) : WearableCommandProcessor {
     override fun init() {
-        commandOutputStream.send(mainResponse {
-            ping = pingResponse { }
-        })
+        commandOutputStream.send(
+            mainResponse {
+                ping = pingResponse { }
+            }
+        )
         commandInputStream.getRequestsFlow().onEach {
             if (it.hasPing()) {
                 commandOutputStream.send(mainResponse { ping = pingResponse { } })