+++ components/wearable/emulate/handheld/impl/src/main/java/com/flipperdevices/wearable/emulate/handheld/impl/service/WearRequestForegroundService.kt
@@ -37,6 +37,7 @@ class WearRequestForegroundService : LifecycleService(), WearRequestChannelBinde
 
         if (intent?.action == CLOSE_CHANNEL_ACTION) {
             stopForeground(STOP_FOREGROUND_REMOVE)
+            stopSelf()
         }
 
         return super.onStartCommand(intent, flags, startId)