+++ components/wearable/emulate/handheld/impl/src/main/java/com/flipperdevices/wearable/emulate/handheld/impl/service/WearRequestListenerService.kt
@@ -59,6 +59,5 @@ class WearRequestListenerService :
     override fun onDestroy() {
         super.onDestroy()
         info { "#onDestroy" }
-        applicationContext.unbindService(this)
     }
 }