+++ components/wearable/emulate/wear/impl/src/main/java/com/flipperdevices/wearable/emulate/impl/viewmodel/FlipperStatusListener.kt
@@ -46,12 +46,13 @@ class FlipperStatusListenerImpl @Inject constructor(
     }
 
     override fun requestStatus() {
-        commandOutputStream.send(mainRequest {
-            subscribeOnConnectStatus = subscribeOnConnectStatusRequest { }
-        })
+        commandOutputStream.send(
+            mainRequest {
+                subscribeOnConnectStatus = subscribeOnConnectStatusRequest { }
+            }
+        )
     }
 
     override fun getState(): StateFlow<ConnectStatusOuterClass.ConnectStatus> =
         connectionStatusStateFlow
-
 }